requirements.txt
django
djangorestframework
transformers
torch